Experiments with optional dyanamic FPS
There are some #defines at the start of the lss servo file that tries to dynamically change the spacing of the messages depending on timing as well as max amount servos move.

Also another option to always output every servo...

Still WIP, I think this delta has it set to same code as before.

Also fixed my hack to know if we are running a PS4 using BT... If the BT object is not active... Then not BT...