Update LSS_WalkAndChewGum.ino
Put time gap in before starting loop, so can see what is output for the stuff.

Also put in some code that tries to show some validation of data.  Not 100% correct
but it also toggles IO pins.
IO 2 - for stuff reading in string, like characters received before *.  Is there a number after the star...

IO 3 - is trying to extract the CMD and number off of the string.  Is there a valid command.  Like < 10 characters before a number...  Then checks to see if a number after this... Probably bugs in this... but does give you cluses.
Then IO4 looks at the command string is it one of the two we asked for...

But does give you clues...